Revert "Tela de pesquisa"
