+++ CineTime/Controller/MyList/MyListViewController.swift
@@ -38,19 +38,10 @@ class MyListViewController: UITableViewController {
         configureNavBar()
         
         // Call Api
-//        fetchData()
+        fetchData()
     }
     
     
-//    func fetchData() {
-//
-//        Service().searchByName(name: "Vingadores") { films in
-//            films?.forEach({ film in
-//                print(film)
-//            })
-//        }
-//
-//    }
     func fetchData() {
         
         Service.shared.searchByName(name: "Vingadores") { films in
