+++ CineTime/Info.plist
@@ -58,9 +58,5 @@
 		<string>UIInterfaceOrientationLandscapeLeft</string>
 		<string>UIInterfaceOrientationLandscapeRight</string>
 	</array>
-	<key>UIAppFonts</key>
-	<array>
-		<string>SeasideResortNF.ttf</string>
-	</array>
 </dict>
 </plist>
