+++ CineTime/Service/Router.swift
@@ -47,8 +47,6 @@ struct FilmsAPI {
             case .nowPlaying:
                 return URL(string: "\(hostname)/movie/now_playing?api_key=\(key)&\(language)%3DBR&page=1")
             case .searchByName(let name):
-                return URL(string: "\(hostname)/search/multi?api_key=\(key)&\(language)&query=\(name)&page=\(page)")
-                //https://api.themoviedb.org/3/search/multi?api_key=<<api_key>>&language=pt-BR&query=entre%20facas%20e%20segredos&page=1&include_adult=false
                 return URL(string: "\(hostname)/search/movie?api_key=\(key)&\(language)&query=\(name)&page=\(page)")
             case .getImageFilm(let name):
                 return URL(string: "\(imageHostnane)/t/p/w500\(name)")
@@ -74,7 +72,7 @@ struct FilmsAPI {
          https://api.themoviedb.org/3/movie/{movie_id}/recommendations?api_key=<<api_key>>&language=en-US&page=1
          
          Example search by name:
-        https://api.themoviedb.org/3/search/multi?api_key=a1d23d9e0cc08e49e4703408659fe47c&language=pt-BR&query=entre%20facas%20e%20segredos&page=1
+         https://api.themoviedb.org/3/search/movie?api_key=a1d23d9e0cc08e49e4703408659fe47c&language=pt-BR&query=Vingadores&page=1
          */
     }
 }
