+++ CineTime/View/Movie.swift
@@ -23,19 +23,6 @@ struct FilmResult: Decodable {
 struct Film: Codable {
     
     let id: Int
-    let title: String?
-    let poster_path: String?
-    let genre_ids: [Int]?
-    let video: Bool = false
-    let overview: String?
-    let release_date: String?
-    let vote_average: Double?
-    let media_type: String?
-
-}
-
-struct SearchResult: Decodable {
-    let id: Int
     let title: String
     let poster_path: String
     let genre_ids: [Int]
@@ -43,7 +30,5 @@ struct SearchResult: Decodable {
     let overview: String
     let release_date: String
     let vote_average: Double
-    let media_type: String
-    
-}
 
+}
