+++ CineTime/View/Search/TitlesCollectionViewCell.swift
@@ -10,14 +10,16 @@ import UIKit
 
 class TitlesCollectionCell: UICollectionViewCell {
 
-
+    var title: Title? {
+        didSet {
+            guard let titleImage = title?.image else {return}
+            
+            imageTitle.image = UIImage(named: titleImage)
+        }
+    }
     override init(frame: CGRect) {
         super.init(frame: frame)
         Setup()
-    }
-    
-    override func prepareForReuse() {
-        imageTitle.image = nil
     }
     
     func Setup() {
@@ -29,7 +31,7 @@ class TitlesCollectionCell: UICollectionViewCell {
     
     var imageTitle: UIImageView = {
         let iv = UIImageView()
-        iv.contentMode = .scaleAspectFill
+        iv.contentMode = .scaleAspectFit
         iv.backgroundColor = .clear
         return iv
     }()
@@ -40,4 +42,70 @@ class TitlesCollectionCell: UICollectionViewCell {
 
 }
 
+struct Title {
+    let image: String?
+}
 
+extension UIView {
+    
+    func setAnchor(top: NSLayoutYAxisAnchor?, left: NSLayoutXAxisAnchor?,
+                bottom: NSLayoutYAxisAnchor?, right: NSLayoutXAxisAnchor?,
+                paddingTop: CGFloat, paddingLeft: CGFloat, paddingBottom: CGFloat,
+                paddingRight: CGFloat, width: CGFloat = 0, height: CGFloat = 0) {
+        
+        self.translatesAutoresizingMaskIntoConstraints = false
+        
+        if let top = top {
+            self.topAnchor.constraint(equalTo: top, constant: paddingTop).isActive = true
+        }
+        
+        if let left = left {
+            self.leftAnchor.constraint(equalTo: left, constant: paddingLeft).isActive = true
+        }
+        
+        if let bottom = bottom {
+            self.bottomAnchor.constraint(equalTo: bottom, constant: paddingBottom).isActive = true
+        }
+        
+        if let right = right {
+            self.rightAnchor.constraint(equalTo: right, constant: -paddingRight).isActive = true
+        }
+        
+        if width != 0 {
+            self.widthAnchor.constraint(equalToConstant: width).isActive = true
+        }
+        
+        if height != 0 {
+            self.heightAnchor.constraint(equalToConstant: height).isActive = true
+        }
+    }
+    
+    var safeTopAnchor: NSLayoutYAxisAnchor {
+        if #available(iOS 11.0, *) {
+            return safeAreaLayoutGuide.topAnchor
+        }
+        return topAnchor
+    }
+    
+    var safeLeftAnchor: NSLayoutXAxisAnchor {
+        if #available(iOS 11.0, *) {
+            return safeAreaLayoutGuide.leftAnchor
+        }
+        return leftAnchor
+    }
+    
+    var safeBottomAnchor: NSLayoutYAxisAnchor {
+        if #available(iOS 11.0, *) {
+            return safeAreaLayoutGuide.bottomAnchor
+        }
+        return bottomAnchor
+    }
+    
+    var safeRightAnchor: NSLayoutXAxisAnchor {
+        if #available(iOS 11.0, *) {
+            return safeAreaLayoutGuide.rightAnchor
+        }
+        return rightAnchor
+    }
+    
+}
