+++ CineTime/View/Storyboard/Base.lproj/LaunchScreen.storyboard
@@ -15,30 +15,13 @@
                     <view key="view" contentMode="scaleToFill" id="Ze5-6b-2t3">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="cineTime" translatesAutoresizingMaskIntoConstraints="NO" id="cbX-3L-hPg">
-                                <rect key="frame" x="70" y="212" width="274" height="472"/>
-                                <constraints>
-                                    <constraint firstAttribute="height" constant="472" id="WqR-UE-ilo"/>
-                                    <constraint firstAttribute="width" constant="274" id="oYp-69-NvL"/>
-                                </constraints>
-                            </imageView>
-                        </subviews>
-                        <color key="backgroundColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <constraints>
-                            <constraint firstItem="cbX-3L-hPg" firstAttribute="centerY" secondItem="Ze5-6b-2t3" secondAttribute="centerY" id="eS7-o0-FI8"/>
-                            <constraint firstItem="cbX-3L-hPg" firstAttribute="centerX" secondItem="Ze5-6b-2t3" secondAttribute="centerX" id="owd-Tz-F9U"/>
-                        </constraints>
                         <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
                         <viewLayoutGuide key="safeArea" id="6Tk-OE-BBY"/>
                     </view>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="iYj-Kq-Ea1" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="52.173913043478265" y="375"/>
+            <point key="canvasLocation" x="53" y="375"/>
         </scene>
     </scenes>
-    <resources>
-        <image name="cineTime" width="399" height="472"/>
-    </resources>
 </document>
